[openwebnet] refactoring to internal package
